Carla-Challenge-dev_pearl/requirements.txt
agents==1.4.0
apex==0.9.10dev
dictor==0.1.12
ephem==4.1.5
gymnasium==0.28.1
huggingface_hub==0.19.4
imgaug==0.4.0
numpy==1.23.5
pexpect==4.9.0
Pillow==10.1.0
py_trees==0.8.3
PyYAML==6.0.1
Shapely==2.0.2
scikit-image==0.21.0
six==1.16.0
simple_watchdog_timer==0.1.1
tabulate==0.9.0
timm==0.9.12
tensorboardX==2.6.2.2
tensorflow==2.13.1
tensorflow_datasets==4.9.2
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
transforms3d==0.4.1
tqdm==4.65.0
wandb==0.16.1
networkx==3.1
